Added migration and models templates Need to create all correct models

+++ .adonisjs/server/routes.d.ts
@@ -1,6 +1,6 @@
 import '@adonisjs/core/types/http'
 
-type ParamValue = string | number | bigint | boolean
+type ParamValue = string | number | integer | boolean
 
 export type ScannedRoutes = {
   ALL: {
